R12 dxf dynamically linked library
This is the initial release of the code for an r12 dxf dynamically
linked library - currently the code supports Windows. The project file
was created under CodeBlocks under R12.11. The code is > 12 years old so
it is in for some significant revision. The actual parser dll is in the
subdirectory named dll.